endret belop fra Int til BigDecimal og totalt reiseAvstand er nutnull

+++ src/test/kotlin/no/nav/tilleggsstonader/sak/util/DomainUtil.kt
@@ -659,7 +659,7 @@ fun vilkårReiseTilSamling(
 fun faktaOffentligTransportReiseTilSamling(
     reiseId: ReiseId = dummyReiseId,
     adresse: String = "Tiltaksveien 1",
-    utgifterOffentligTransport: Int = 40,
+    utgifterOffentligTransport: BigDecimal = 40.toBigDecimal(),
 ) = FaktaOffentligTransportReiseTilSamling(
     reiseId = reiseId,
     adresse = adresse,
